Completed Lesson 2.2 - Food Flight
Created MoveForward script and attached it to the food object.

Added speed variable and a transform.Translate statement for food projectile flight.

Created a Prefab folder and added food object as an projectile prefab.

Added a GameObject to PlayerController script for our food projectile.

Added an Instantiate operation that launches a food projectile on every press of the space bar.

Changed the speed of and created prefabs for each animal.

Created DestroyOutOfBounds script and applied it to food objects as an override so they are destroyed when flying outside of the scene boundaries.

Added else if statement for destroying animal gameObjects when they go outside of the scene bounds.