Completed Lesson 3.2 - Make the World Whiz By
Created a script to loop the background seamlessly.

Modified the PlayerController script to end the game on collision with an obstacle but not on a collision with the ground.

Modified the MoveLeft script to stop obstacles and background from moving once the player triggers gameOver.

Modified SpawnManager script to stop spawning obstacles once the player has triggered gameOver.

Modified the MoveLeft script so obstacles that exit the scene boundaries are destroyed.