Completed Lesson 3.3 - Don't Just Stand There
Became familiar with the Player's Animation states and transition requirements.

Modified Player's Speed_f to 1 to achieve the run_static animation conditions, increased the run_static speed to match the Player's movement with the speed of the scrolling background, and set the default entry state to run_static so that the player is running from the start of the game.

Modified the Running_Jump state to slow the animation down to match the other speeds in the scene and modified the player's RigidBody mass, jumpForce, and gravity values for better gameplay.

Added a death state animation to the player and set it to run on game over.

Added a condition to the PlayerController Update method to prevent the player from jumping if the game is over.