Completed Lesson 4.1 - Watch Where You're Going.
Added player object and modified its physical attributes.

Created a focal point and attached the main camera as a child object to have it pinned to the center of the island and rotate around the game.

Created Focal Point and Player Controller scripts to control camera and player movement via user input.