Completed Lesson 4.2 - Follow the Player
Created an enemy sphere and modified its physical attributes.

Created a new bouncy physics material and applied it to the player and enemy objects.

Created enemy script to make enemy spheres move in the direction of the player.

Created a spawn manager object to spawn enemies in a random position.

Improved code and created a method within the Enemy script for random enemy spawn location.