Completed Challenge 1 - Plane Programming
Modified PlayerControllerX script by changing Vector3.back to Vector3.forward to make the plane go forward.

Slowed the plane down by adding Time.deltaTime to the plane speed calculation.

Prevented the plane from tilting without user input by adding the verticalInput value to the rotation calculation.

Repositioned the camera to be beside the plane and to follow the player.

Created a SpinPropellerX script and applied it to the Propeller object to spin the propeller around its Z axis for simulating normal propeller spin.