Completed Challenge 2 - Play Fetch Extra Credit
Added three ball prefabs as the spawn manager script elements so the balls spawn at the top of the screen.

Changed the Dog Prefab in the PlayerControllerX script to Dog from Ball to make the player spawn dogs.

Modified the dog object box collider so balls are only destroyed upon coming into contact.

Corrected the lowerLimit variable to be a negative value, changed > to <, and changed y to z in the DestroyOutofBounds script.

Added a random index variable to randomize the balls being spawned.