Completed Lesson 4.3 - PowerUp and CountDown
Added a powerup object and modified its physical properties.

Modified the PlayerController script to destroy power ups on contact with the player.

Modified the PlayerController script to test for collisions with enemies when the player has a powerup.

Modified the PlayerController script to apply extra force to enemy collisions when the player has collected a power up.

Added a coroutine which implements a countdown during which the power up will remain active.

Added a visual power up indicator to show when the player does or does not have the power up ability.