Completed Challenge 3 - Balloons, Bombs, & Booleans Extra Credit
Modified the player controller script by assigning the player's rigid body variable so the user can control the balloon movement.

Modified the MoveLeft script so that game objects move left only if the game is not over.

Corrected the name of the SpawnManagerX.SpawnsObject so game objects spawn.

Corrected the background repeatWidth variable so the background repeats smoothly.

Added a boolean to make sure the player doesn't go too high.

Added a boolean to detect if the balloon is on the ground and if so the balloon has an upwards force applied to simulate bouncing.

Corrected location of the fireworks to next to the balloon.