Completed Lesson 5.2 - Keeping Score
Added a TextMeshPro object and anchored it to the top left corner of the canvas in preparation for creating a game scoreboard.

Change the TextMeshPro text to Score:, modified the text font, style, and size.

Modified the GameManager script to set the scoreboard text to include a score variable.

Modified the GameManager script by adding an UpdateScore method that will update as the player plays the game to keep track of their score.

Modified GameManager and Target scripts to increase the player's score by 5 every time they destroy one of the game objets.

Modified Target script by adding a new pointValue variable and assigning our game objects different point values including a negative value for the bad object.

Added an explosionParticle variable to the Target script and assigned each prefab a different color explosion which will show when the objects are clicked and destroyed.