Completed Lesson 5.3 - Game Over
Added a TextMeshPro object to the canvas and modified its attributes for placement and style.

Modified the GameManager and Target scripts and added a tag to the bad object to show the game over text only if a good object falls out of the scene.

Modified both scripts by adding a condition for checking that the game is active, and if the game is not active stops objects from spawning.

Added a Restart UI button to allow the player to play the game again after losing.

Added Restart button and modified its font, size, and color.

Modified the GameManager script by adding a RestartGame method and attaching it to our Restart button and set it to restart the game when clicked.

Modified GameManager script to only show the Restart button when the game is not active.