Completed Lesson 5.4 - What's the Difficulty?
Added and stylized a game title and Easy, Medium, and Hard difficulty buttons.

Created a difficulty button script and attached to all three difficulty buttons.

Modified the difficulty button script by adding a new SetDifficulty method which will write which button was clicked to the console.

Modified the GameManager and DifficultyButton scripts to make the difficulty buttons start the game.

Moved title text and buttons to a new empty Title Screen object and modified the GameManager script to make the object inactive when the game starts.

Modified the DifficultyButton and GameManager scripts to add a variable and calculation to set the game difficulty and assigned values from 1-3 to the buttons respectively.